+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [v0.3.3] - 2023-03-03
 
 ### Added
 
@@ -62,6 +62,7 @@ The two most important changes in usage are:
 [0.3.0]: https://github.com/dogmatiq/ferrite/releases/tag/v0.3.0
 [0.3.1]: https://github.com/dogmatiq/ferrite/releases/tag/v0.3.1
 [0.3.2]: https://github.com/dogmatiq/ferrite/releases/tag/v0.3.2
+[0.3.3]: https://github.com/dogmatiq/ferrite/releases/tag/v0.3.3
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD